Move redux structure to redux folder
