+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './styles/index.css';
 import { Provider } from 'react-redux';
 import App from './containers/App';
-import store from './store';
+import store from './redux/store';
 
 ReactDOM.render(
   <React.StrictMode>
